refactor(user): Phase 3 — SocialAccount→UserIdentity + AuthProvider 확장 + UserProfile Integer + 의존 서비스
DBv5 정합. SocialAccount/SocialAccountRepository 폐기, 신규 UserIdentity
(테이블 user_identities, PK 컬럼 id Integer, created_at, email 컬럼 추가,
UNIQUE(user_id,provider) 제거 — DBv5 정합) + UserIdentityRepository.

AuthProvider enum 확장: KAKAO + GOOGLE/NAVER/APPLE (DBv5 CHECK). LOCAL
은 응답 표기 전용 — DB CHECK 에 없으며 user_identities row 없음으로 판정.

AuthService.signUp/login: request.getUsername() / existsByUsername /
findByUsername / User.builder().username(...) / Integer userId 통일.
SocialAuthService: UserIdentityRepository 사용, Integer userId,
user_identities.email 컬럼에 OAuth 이메일 별도 저장.

UserProfile / UserProfileRepository: userId Long → Integer,
recentRecipeIds List<Long> → List<Integer>, empty(Integer).

본 commit 단독으론 UserMeService/Controller/AdminUserService 등이 옛
시그니처(Long userId) 호출하고 있어 compile fail. Phase 4 (UserRecipe rename
+ Like/Scrap/UserMeService/Controller Integer ripple) 후 정합.

+++ src/main/java/com/naengo/api_server/domain/user/entity/UserProfile.java
@@ -35,7 +35,7 @@ public class UserProfile {
 
     @Id
     @Column(name = "user_id")
-    private Long userId;
+    private Integer userId;
 
     @JdbcTypeCode(SqlTypes.JSON)
     @Column(name = "user_input", columnDefinition = "jsonb", nullable = false)
@@ -82,7 +82,7 @@ public class UserProfile {
 
     @JdbcTypeCode(SqlTypes.JSON)
     @Column(name = "recent_recipe_ids", columnDefinition = "jsonb")
-    private List<Long> recentRecipeIds;
+    private List<Integer> recentRecipeIds;
 
     @Column(name = "ai_analyzed_at")
     private ZonedDateTime aiAnalyzedAt;
@@ -91,7 +91,7 @@ public class UserProfile {
     private ZonedDateTime updatedAt;
 
     /** 빈 default — 신규 사용자가 처음 PUT 할 때 사용. updated_at NOT NULL 충족. */
-    public static UserProfile empty(Long userId) {
+    public static UserProfile empty(Integer userId) {
         return UserProfile.builder()
                 .userId(userId)
                 .userInput(List.of())
